always update from a required node
previous behavior would leave a running cluster alone even if it matched
a required edge, and only move it if it was changed.

this led to a couple of edge cases where it was possible to not detect
required edges. instead, if the current state matches a required edge,
the update is always taken.

also adds helpers for building update graphs