Combines `CtxCluster` and `CtxClusterStatus` into one.
The original idea was that `CtxCluster` was a reference to the `SpiceDBCluster` in the informer cache (which shouldn't be modified) so any handler could recover the "original" object being synced if it needed it. And the `CtxClusterStatus` was a copy of the `SpiceDBCluster` object that would be continually updated if we made changes to it during a sync, so that subsequent handlers would get the "new" view of the object.

But in practice it's hard to remember the distinction, so this commit
changes `CtxCluster` to be a deepcopy of the object from the cache, and it's easy enough to query the cache directly if we need to recover the original object in the future.